Move Windows intergration test.
